# SQS-406 | E2E Tests: Increase expected latency upper bound for /passthrough/active-orders enpoind ( backport )