This PR Increases latency upper bound from 150ms to 450ms based on our current observations via monitoring tools to make tests less noisy.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **Bug Fixes**
	- Adjusted the expected latency upper bound to better reflect performance capabilities, increasing from 150ms to 450ms. This change may enhance testing accuracy and align with updated performance benchmarks.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->